requirements.txt
fastapi
uvicorn
pydantic
langchain
langgraph
langchain-ollama
python-dotenv
